requirements.txt
cftime==1.4.1
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
netCDF4==1.5.6
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
